+++ README.md
@@ -570,6 +570,53 @@ func main() {
 }
 ```
 
+### TinyGo Support through MCU UART Transport
+
+```go
+//go:build baremetal
+
+package main
+
+import (
+    "context"
+    "log"
+
+    "github.com/hipsterbrown/feetech-servo/feetech"
+)
+
+func main() {
+    ctx := context.Background()
+
+    // Create a new servo bus
+    bus, err := feetech.NewBus(feetech.BusConfig{
+        Port:     "0",
+        BaudRate: 1000000,
+        Protocol: feetech.ProtocolSTS,
+    })
+    if err != nil {
+        log.Fatal("Failed to create bus:", err)
+    }
+    defer bus.Close()
+
+    // Create a servo instance (defaults to STS3215)
+    servo := feetech.NewServo(bus, 1, nil)
+
+    // Detect model
+    if err := servo.DetectModel(ctx); err != nil {
+        log.Fatal("Failed to detect model:", err)
+    }
+    log.Printf("Connected to: %s", servo.Model().Name)
+
+    // Enable torque and move to center position
+    servo.Enable(ctx)
+    servo.SetPosition(ctx, 2048) // Center position for 12-bit servo
+
+    // Read current position
+    pos, _ := servo.Position(ctx)
+    log.Printf("Current position: %d", pos)
+}
+```
+
 ### Testing with MockTransport
 
 ```go
